Altered to try it out with urdflbriiwa14 package

+++ include/ros2_cortex/MarkerPublisher.hpp
@@ -20,7 +20,8 @@ private:
      rclcpp_lifecycle::LifecyclePublisher<visualization_msgs::msg::MarkerArray>::SharedPtr marker_array_publisher_;
      visualization_msgs::msg::MarkerArray marker_array_;
      float frame_delta_time_ns_ = -1.0;
-     float nss_in_s = 1000000000;
+//     float nss_in_s = 1000000000;
+     int nss_in_s = 1000000000;
      std::string fps_comm = "GetContextFrameRate";
      rclcpp::QoS qos;
 };
